# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    API specification. Requests to an unknown route under `/api` return a 404
    response using the `Error` schema with the message `API route not found`.
    Requests with malformed JSON bodies return a 400 response using the `Error`
    schema with the message `Malformed JSON request body`. Unexpected server
    failures return a generic 500 response using the same schema. JSON request
    bodies that exceed the parser's size limit return a 413 response using the
    `Error` schema with the message `Request body too large`.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start browser login
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the identity provider
...
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: End the browser session
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the identity provider
...
  /instructor/preferences:
    get:
      operationId: getInstructorPreferences
      tags: [instructor]
      summary: Get the signed-in instructor's workspace preferences
      responses:
        "200":
          description: Instructor workspace preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorPreferences"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Instructor role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateInstructorPreferences
      tags: [instructor]
      summary: Update the signed-in instructor's workspace preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstructorPreferencesUpdate"
      responses:
        "200":
          description: Updated instructor workspace preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorPreferences"
        "400":
          description: Invalid instructor preference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Instructor role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /bookmarks/questions:
    get:
      operationId: listBookmarkQuestions
      tags: [practice]
      summary: List the learner's saved questions
      responses:
        "200":
          description: Saved learner questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "400":
          description: Learner identity is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bookmarks/{questionId}:
    put:
      operationId: updateBookmark
      tags: [practice]
      summary: Save or remove a learner bookmark
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "200":
          description: Updated bookmark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "400":
          description: Invalid bookmark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: JSON request body exceeds the size limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
...
  schemas:
...
    InstructorPreferences:
      type: object
      required: [pageSize]
      properties:
        pageSize:
          type: integer
          enum: [10, 25, 50]
    InstructorPreferencesUpdate:
      type: object
      required: [pageSize]
      properties:
        pageSize:
          type: integer
          enum: [10, 25, 50]
...
    QuestionDomain:
      type: string
      enum: [People, Process, "Business environment"]
    QuestionApproach:
      type: string
      enum: [agile, predictive, hybrid]
    QuestionDifficulty:
      type: string
      enum: [easy, medium, hard]
    QuestionStatus:
      type: string
      enum: [draft, published, archived]
    Question:
      type: object
      required:
        - id
        - domain
        - topic
        - approach
        - difficulty
        - question
        - translation
        - options
        - correctAnswer
        - explanation
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        domain:
          $ref: "#/components/schemas/QuestionDomain"
        topic:
          type: string
          minLength: 1
        approach:
          $ref: "#/components/schemas/QuestionApproach"
        difficulty:
          $ref: "#/components/schemas/QuestionDifficulty"
        question:
          type: string
          minLength: 1
        translation:
          type: string
        options:
          type: array
          minItems: 2
          maxItems: 6
          items:
            type: string
            minLength: 1
        correctAnswer:
          type: integer
          minimum: 0
        explanation:
          type: string
          minLength: 1
        status:
          $ref: "#/components/schemas/QuestionStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BookmarkInput:
      type: object
      required: [saved]
      properties:
        saved:
          type: boolean
    Bookmark:
      type: object
      required: [questionId, saved]
      properties:
        questionId:
          type: string
        saved:
          type: boolean
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string