# lib/api-zod/src/generated/types/questionDifficulty.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification. Requests to an unknown route under `/api` return a 404
 * response using the `Error` schema with the message `API route not found`.
 * Requests with malformed JSON bodies return a 400 response using the `Error`
 * schema with the message `Malformed JSON request body`. Unexpected server
 * failures return a generic 500 response using the same schema. JSON request
 * bodies that exceed the parser's size limit return a 413 response using the
 * `Error` schema with the message `Request body too large`.
 *
 * OpenAPI spec version: 0.1.0
 */

export type QuestionDifficulty = typeof QuestionDifficulty[keyof typeof QuestionDifficulty];


export const QuestionDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;